Add changes to app.js. Remove controllers/ folder. Add routes/ folder and router file.
