--- controllers/donationController.js
@@ -1,9 +1,0 @@
-function donationController() {
-  return {
-    renderDonations: function* (){
-      yield this.render('donations', {});
-    }
-  };
-}
-
-module.exports = donationController;
