--- controllers/eventController.js
@@ -1,9 +1,0 @@
-function eventController() {
-  return {
-    renderEvents: function* (){
-      yield this.render('events', {});
-    }
-  };
-}
-
-module.exports = eventController;
