+++ package.json
@@ -15,26 +15,26 @@
     "del": "^1.2.0",
     "gulp": "~3.9.0",
     "gulp-autoprefixer": "^2.3.1",
+    "gulp-changed": "^1.3.0",
     "gulp-minify-css": "^1.2.0",
     "gulp-nodemon": "^2.0.6",
     "gulp-notify": "^2.2.0",
     "gulp-plumber": "^1.0.1",
     "gulp-rename": "^1.2.2",
     "gulp-sass": "^2.0.4",
+    "gulp-shell": "^0.5.2",
     "kerberos": "0.0.17",
     "koa": "1.1.2",
     "koa-livereload": "^0.1.23",
-    "koa-route": "^2.4.2",
+    "koa-router": "^5.4.0",
     "koa-static": "^1.5.2",
     "koa-views": "^3.1.0",
+    "migrate": "^0.2.2",
     "mongodb": "^1.4.39",
     "mongoose": "^4.3.4",
     "monk": "^1.0.1",
     "nunjucks": "^2.2.0",
-    "validator": "^4.4.0",
-    "gulp-shell": "^0.5.2",
-    "gulp-changed": "^1.3.0",
-    "migrate": "^0.2.2"
+    "validator": "^4.4.0"
   },
   "engines": {
     "node": "5.0.0"
