+++ routes/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  donation: require('./donation').routes(),
+  event: require('./event').routes(),
+  home: require('./home').routes(),
+  talk: require('./talk').routes(),
+  user: require('./user').routes()
+};
